Add files via upload
In this project, I transferred the Real Estate dataset into SAS Studio for a more in-depth analysis. SAS Studio, a powerful statistical analysis tool, was employed to conduct a comprehensive series of Regression Analyses. This involved scrutinizing various combinations of independent variables to identify a more refined set that could better explain changes in the dependent variable (Price). The goal was to enhance the accuracy and reliability of the regression model by selecting the most influential explanatory variables.